Update status, remove user_role

+++ app/dto/UserResponse.go
@@ -4,5 +4,5 @@ type UserResponse struct {
 	UserId   uint   `json:"user_id"`
 	Name     string `json:"name"`
 	Username string `json:"username"`
-	Gender   int8   `json:"gender"`
+	Gender   int    `json:"gender"`
 }
