+++ app/model/Permission.go
@@ -3,5 +3,5 @@ package model
 type Permission struct {
 	PermissionId uint   `gorm:"primaryKey" json:"permission_id"`
 	Name         string `gorm:"not null;unique" json:"name"`
-	Status       int8   `gorm:"default:1" json:"status"`
+	Status       int    `gorm:"default:1" json:"status"`
 }
